+++ docs/examples/button/roleButtonExample.js
@@ -5,7 +5,7 @@ import { Button, Flex } from 'gestalt';
 export default function Example(): Node {
   return (
     <Flex alignItems="center" height="100%" justifyContent="center" width="100%">
-      <Button accessibilityLabel="Follow" size="lg" text="Follow" role="button" />
+      <Button accessibilityLabel="Follow" size="lg" text="Follow" />
     </Flex>
   );
 }
