+++ docs/examples/popover/variantLayer.js
@@ -189,13 +189,7 @@ export default function Example(): Node {
                 </Flex.Item>
                 <Flex gap={{ column: 0, row: 2 }}>
                   <Button text="Cancel" size="lg" onClick={() => setShowSheet(false)} />
-                  <Button
-                    text="Done"
-                    color="red"
-                    size="lg"
-                    type="submit"
-                    onClick={() => setShowSheet(false)}
-                  />
+                  <Button text="Done" color="red" size="lg"  onClick={() => setShowSheet(false)} />
                 </Flex>
               </Flex>
             }
