+++ packages/eslint-plugin-gestalt/src/__fixtures__/button-icon-restrictions/invalid/invalid-wrong-link-role-icon.js
@@ -1,5 +1,5 @@
 import { Button } from 'gestalt';
 
 export default function TestElement() {
-  return <Button color="white" iconEnd="arrow-up-right" size="lg" text="Link" role="link" />;
+  return <Button color="white" iconEnd="arrow-up-right" size="lg" text="Link"  />;
 }
