+++ packages/eslint-plugin-gestalt/src/__fixtures__/button-icon-restrictions/valid/valid-size.js
@@ -2,9 +2,9 @@ import { Button } from 'gestalt';
 
 export default function TestElement() {
   return (
-  <Box>
-    <Button color="white" iconEnd="arrow-down" size="lg" text="Menu" />
-    <Button role="link" iconEnd="visit" size="lg" text="link" />
-  </Box>
+    <Box>
+      <Button color="white" iconEnd="arrow-down" size="lg" text="Menu" />
+      <Button  iconEnd="visit" size="lg" text="link" />
+    </Box>
   );
 }
