+++ packages/eslint-plugin-gestalt/src/__fixtures__/no-role-link-components/invalid/invalid-button.js
@@ -1,5 +1,5 @@
 import { Button } from 'gestalt';
 
 export default function TestElement() {
-  return <Button role='link'/>
+  return <Button  />;
 }
