+++ packages/gestalt/src/ActivationCard.js
@@ -3,7 +3,7 @@ import { Fragment, type Node } from 'react';
 import classnames from 'classnames';
 import styles from './ActivationCard.css';
 import Box from './Box.js';
-import Button from './Button/Button.js';
+import ButtonLink from './ButtonLink.js';
 import { useDefaultLabelContext } from './contexts/DefaultLabelProvider.js';
 import Icon from './Icon.js';
 import IconButton from './IconButton.js';
@@ -21,11 +21,7 @@ type LinkData = {|
   href: string,
   label: string,
   onClick?: ({|
-    event:
-      | SyntheticMouseEvent<HTMLButtonElement>
-      | SyntheticMouseEvent<HTMLAnchorElement>
-      | SyntheticKeyboardEvent<HTMLAnchorElement>
-      | SyntheticKeyboardEvent<HTMLButtonElement>,
+    event: SyntheticMouseEvent<HTMLAnchorElement> | SyntheticKeyboardEvent<HTMLAnchorElement>,
     dangerouslyDisableOnNavigation: () => void,
   |}) => void,
   rel?: 'none' | 'nofollow',
@@ -85,14 +81,13 @@ function ActivationCardLink({ data }: {| data: LinkData |}): Node {
       marginStart="auto"
       rounding="pill"
     >
-      <Button
+      <ButtonLink
         accessibilityLabel={accessibilityLabel}
         color="gray"
         href={href}
         fullWidth
         onClick={onClick}
         rel={rel}
-        role="link"
         size="lg"
         text={label}
         target={target}
