+++ packages/gestalt/src/Button/Button.js
@@ -1,48 +1,37 @@
 // @flow strict
-import {
-  type AbstractComponent,
-  forwardRef,
-  Fragment,
-  type Node,
-  useImperativeHandle,
-  useRef,
-} from 'react';
-import classnames from 'classnames';
-import styles from './Button.css';
-import getAriaLabel from '../accessibility/getAriaLabel.js';
-import NewTabAccessibilityLabel from '../accessibility/NewTabAccessibilityLabel.js';
-import { useColorScheme } from '../contexts/ColorSchemeProvider.js';
-import { useDefaultLabelContext } from '../contexts/DefaultLabelProvider.js';
-import Flex from '../Flex.js';
-import focusStyles from '../Focus.css';
-import Icon, { type IconColor } from '../Icon.js';
+import { type AbstractComponent, forwardRef, type Node } from 'react';
+import InternalButton from './InternalButton.js';
 import icons from '../icons/index.js';
-import InternalLink from '../Link/InternalLink.js';
-import touchableStyles from '../TapArea.css';
-import Text from '../Text.js';
-import useFocusVisible from '../useFocusVisible.js';
-import useTapFeedback from '../useTapFeedback.js';
 
-const DEFAULT_TEXT_COLORS = {
-  blue: 'inverse',
-  gray: 'default',
-  red: 'inverse',
-  transparent: 'default',
-  semiTransparentWhite: 'default',
-  transparentWhiteText: 'inverse',
-  white: 'default',
-};
-
-const SIZE_NAME_TO_PIXEL = {
-  sm: 10,
-  md: 12,
-  lg: 12,
-};
-
-type Target = null | 'self' | 'blank';
-
-type BaseButton = {|
+type Props = {|
+  /**
+   * Text to render inside the Button to convey the function and purpose of the Button.
+   */
+  text: string,
+  /**
+   * A unique id indicating the element or elements whose contents or visibility are controlled by Button.
+   * See the [Accessibility guidelines](https://gestalt.pinterest.systems/foundations/accessibility) to learn more.
+   */
+  accessibilityControls?: string,
+  /**
+   * Needed if Button controls the visibility of other elements, e.g. Dropdown or Flyout.
+   * This is used to indicate if the controlled grouping is currently expanded or collapsed.
+   * See the [Accessibility guidelines](https://gestalt.pinterest.systems/foundations/accessibility) to learn more.
+   */
+  accessibilityExpanded?: boolean,
+  /**
+   * Set as true if Button controls one or more interactive popup elements, such as a menu or dialog.
+   * See the [Accessibility guidelines](https://gestalt.pinterest.systems/foundations/accessibility) to learn more.
+   */
+  accessibilityHaspopup?: boolean,
+  /**
+   * Label to provide more context around Button’s function or purpose.
+   * See the [Accessibility guidelines](https://gestalt.pinterest.systems/foundations/accessibility) to learn more.
+   */
   accessibilityLabel?: string,
+  /**
+   * The background color of Button.
+   */
   color?:
     | 'gray'
     | 'red'
@@ -51,83 +40,49 @@ type BaseButton = {|
     | 'semiTransparentWhite'
     | 'transparentWhiteText'
     | 'white',
+  /**
+   * Available for testing purposes, if needed. Consider [better queries](https://testing-library.com/docs/queries/about/#priority) before using this prop.
+   */
   dataTestId?: string,
+  /**
+   * Indicates if Button is disabled. Disabled Buttons are inactive and cannot be interacted with.
+   */
   disabled?: boolean,
-  iconEnd?: $Keys<typeof icons>,
+  /**
+   * Default Buttons are sized by the text within the Button whereas full-width Buttons expand to the full width of their container.
+   */
   fullWidth?: boolean,
+  /**
+   * An icon displayed after the text to help clarify the usage of Button. See the [icon variant](https://gestalt.pinterest.systems/web/button#Icons) to learn more.
+   */
+  iconEnd?: $Keys<typeof icons>,
+  /**
+   * The name attribute specifies the name of the \<button\> element.
+   * The name attribute is used to reference form-data after the form has been submitted.
+   */
+  name?: string,
+  /**
+   * Callback invoked when the user clicks (press and release) on Button with the mouse or keyboard. Required with `role="button"` or `type="button"` Buttons.
+   * See [GlobalEventsHandlerProvider](/web/utilities/globaleventshandlerprovider#Link-handlers) to learn more about link navigation.,
+   */
   onClick?: ({|
-    event:
-      | SyntheticMouseEvent<HTMLButtonElement>
-      | SyntheticMouseEvent<HTMLAnchorElement>
-      | SyntheticKeyboardEvent<HTMLAnchorElement>
-      | SyntheticKeyboardEvent<HTMLButtonElement>,
+    event: SyntheticMouseEvent<HTMLButtonElement> | SyntheticKeyboardEvent<HTMLButtonElement>,
     dangerouslyDisableOnNavigation: () => void,
   |}) => void,
-  tabIndex?: -1 | 0,
-  size?: 'sm' | 'md' | 'lg',
-  text: string,
-|};
-
-type ButtonType = {|
-  ...BaseButton,
-  accessibilityControls?: string,
-  accessibilityExpanded?: boolean,
-  accessibilityHaspopup?: boolean,
+  /**
+   * Indicates if Button is currently selected.
+   */
   selected?: boolean,
-  type?: 'button',
-  role?: 'button',
-  name?: string,
+  /**
+   * sm: 32px, md: 40px, lg: 48px
+   */
+  size?: 'sm' | 'md' | 'lg',
+  /**
+   * Use "-1" to remove Button from keyboard navigation.
+   * See the [Accessibility guidelines](https://gestalt.pinterest.systems/foundations/accessibility) to learn more.
+   */
+  tabIndex?: -1 | 0,
 |};
-
-type SubmitButtonType = {|
-  ...BaseButton,
-  type: 'submit',
-  role?: 'button',
-  name?: string,
-|};
-
-type LinkButtonType = {|
-  ...BaseButton,
-  href: string,
-  rel?: 'none' | 'nofollow',
-  role: 'link',
-  target?: Target,
-|};
-
-type unionProps = ButtonType | SubmitButtonType | LinkButtonType;
-
-type unionRefs = HTMLButtonElement | HTMLAnchorElement;
-
-function InternalButtonContent({
-  target,
-  text,
-  textColor,
-  icon,
-  size,
-}: {|
-  target?: Target,
-  text: Node,
-  textColor: IconColor,
-  icon?: $Keys<typeof icons>,
-  size: string,
-|}): Node {
-  return (
-    <Fragment>
-      <Flex alignItems="center" gap={{ row: 2, column: 0 }} justifyContent="center">
-        {text}
-        {icon ? (
-          <Icon
-            accessibilityLabel=""
-            color={textColor}
-            icon={icon}
-            size={SIZE_NAME_TO_PIXEL[size]}
-          />
-        ) : null}
-      </Flex>
-      <NewTabAccessibilityLabel target={target} />
-    </Fragment>
-  );
-}
 
 /**
  * [Buttons](https://gestalt.pinterest.systems/web/button) allow users to perform actions within a surface. They can be used alone for immediate action, or as a trigger for another component, like [Dropdown](https://gestalt.pinterest.systems/web/dropdown) or [Popover](https://gestalt.pinterest.systems/web/popover).
@@ -136,212 +91,49 @@ function InternalButtonContent({
  * ![Button dark mode](https://raw.githubusercontent.com/pinterest/gestalt/master/playwright/visual-test/Button-dark.spec.mjs-snapshots/Button-dark-chromium-darwin.png)
  *
  */
-const ButtonWithForwardRef: AbstractComponent<unionProps, unionRefs> = forwardRef<
-  unionProps,
-  unionRefs,
->(function Button(props: unionProps, ref): Node {
-  const {
+const ButtonWithForwardRef: AbstractComponent<Props, HTMLButtonElement> = forwardRef<
+  Props,
+  HTMLButtonElement,
+>(function Button(
+  {
+    text,
+    accessibilityControls,
+    accessibilityExpanded,
+    accessibilityHaspopup,
     accessibilityLabel,
     color = 'gray',
     dataTestId,
     disabled = false,
     fullWidth = false,
     iconEnd,
+    name,
     onClick,
-    tabIndex = 0,
     selected = false,
     size = 'md',
-    text,
-  } = props;
-
-  const innerRef = useRef<null | HTMLAnchorElement | HTMLButtonElement>(null);
-
-  // When using both forwardRef and innerRef, React.useimperativehandle() allows a parent component
-  // that renders <Button ref={inputRef} /> to call inputRef.current.focus()
-  useImperativeHandle(ref, () => innerRef.current);
-
-  const {
-    compressStyle,
-    isTapping,
-    handleBlur,
-    handleMouseDown,
-    handleMouseUp,
-    handleTouchStart,
-    handleTouchMove,
-    handleTouchCancel,
-    handleTouchEnd,
-  } = useTapFeedback({
-    height: innerRef?.current?.clientHeight,
-    width: innerRef?.current?.clientWidth,
-  });
-
-  const { accessibilityNewTabLabel } = useDefaultLabelContext('Link');
-
-  const { name: colorSchemeName } = useColorScheme();
-  // We need to make a few exceptions for accessibility reasons in darkMode for red buttons
-  const isDarkMode = colorSchemeName === 'darkMode';
-  const isDarkModeRed = isDarkMode && color === 'red';
-  const isDarkModeBlue = isDarkMode && color === 'blue';
-
-  let colorClass = color === 'transparentWhiteText' ? 'transparent' : color;
-
-  if (isDarkModeRed) {
-    colorClass = 'darkModeRed';
-  }
-
-  const { isFocusVisible } = useFocusVisible();
-
-  const sharedTypeClasses = classnames(styles.button, focusStyles.hideOutline, {
-    [styles.inline]: props.role !== 'link' && !fullWidth,
-    [styles.block]: props.role !== 'link' && fullWidth,
-    [focusStyles.accessibilityOutline]: !disabled && isFocusVisible,
-  });
-
-  const baseTypeClasses = classnames(sharedTypeClasses, touchableStyles.tapTransition, {
-    [styles.sm]: size === 'sm',
-    [styles.md]: size === 'md',
-    [styles.lg]: size === 'lg',
-    // $FlowFixMe[invalid-computed-prop]
-    [styles[colorClass]]: !disabled && !selected,
-    [styles.selected]: !disabled && selected,
-    [styles.disabled]: disabled,
-    [styles.enabled]: !disabled,
-    [touchableStyles.tapCompress]: props.role !== 'link' && !disabled && isTapping,
-  });
-
-  const parentButtonClasses = classnames(sharedTypeClasses, styles.parentButton);
-
-  const childrenDivClasses = classnames(baseTypeClasses, styles.childrenDiv);
-
-  const textColor =
-    (disabled && 'subtle') ||
-    (selected && 'inverse') ||
-    ((isDarkModeRed || isDarkModeBlue) && 'default') ||
-    DEFAULT_TEXT_COLORS[color];
-
-  const buttonText = (
-    <Text align="center" color={textColor} overflow="normal" weight="bold">
-      {text}
-    </Text>
-  );
-
-  const handleClick = (
-    event: SyntheticKeyboardEvent<HTMLAnchorElement> | SyntheticMouseEvent<HTMLAnchorElement>,
-    dangerouslyDisableOnNavigation: () => void,
-  ) =>
-    onClick
-      ? onClick({
-          event,
-          dangerouslyDisableOnNavigation: dangerouslyDisableOnNavigation ?? (() => {}),
-        })
-      : undefined;
-
-  const handleLinkClick = ({
-    event,
-    dangerouslyDisableOnNavigation,
-  }: {|
-    dangerouslyDisableOnNavigation: () => void,
-    event: SyntheticMouseEvent<HTMLAnchorElement> | SyntheticKeyboardEvent<HTMLAnchorElement>,
-  |}) => handleClick(event, dangerouslyDisableOnNavigation);
-
-  if (props.role === 'link') {
-    const { href, rel = 'none', target = null } = props;
-
-    const ariaLabel = getAriaLabel({ target, accessibilityLabel, accessibilityNewTabLabel });
-
-    return (
-      <InternalLink
-        accessibilityLabel={ariaLabel}
-        colorClass={colorClass}
-        dataTestId={dataTestId}
-        disabled={disabled}
-        fullWidth={fullWidth}
-        href={href}
-        onClick={handleLinkClick}
-        ref={innerRef}
-        rel={rel}
-        tabIndex={tabIndex}
-        selected={selected}
-        size={size}
-        target={target}
-        wrappedComponent="button"
-      >
-        <InternalButtonContent
-          target={target}
-          text={buttonText}
-          textColor={textColor}
-          icon={iconEnd}
-          size={size}
-        />
-      </InternalLink>
-    );
-  }
-
-  if (props.type === 'submit') {
-    const { name } = props;
-
-    return (
-      <button
-        aria-label={accessibilityLabel}
-        className={baseTypeClasses}
-        data-test-id={dataTestId}
-        disabled={disabled}
-        name={name}
-        onBlur={handleBlur}
-        onClick={handleClick}
-        onMouseDown={handleMouseDown}
-        onMouseUp={handleMouseUp}
-        onTouchCancel={handleTouchCancel}
-        onTouchEnd={handleTouchEnd}
-        onTouchMove={handleTouchMove}
-        onTouchStart={handleTouchStart}
-        ref={innerRef}
-        style={compressStyle || undefined}
-        tabIndex={disabled ? null : tabIndex}
-        type="submit"
-      >
-        <InternalButtonContent text={buttonText} textColor={textColor} icon={iconEnd} size={size} />
-      </button>
-    );
-  }
-
-  const { accessibilityControls, accessibilityExpanded, accessibilityHaspopup, name } = props;
-
+    tabIndex = 0,
+  }: Props,
+  ref,
+): null | Node {
   return (
-    <button
-      aria-controls={accessibilityControls}
-      aria-expanded={accessibilityExpanded}
-      aria-haspopup={accessibilityHaspopup}
-      aria-label={accessibilityLabel}
-      className={parentButtonClasses}
-      data-test-id={dataTestId}
+    <InternalButton
+      accessibilityControls={accessibilityControls}
+      accessibilityExpanded={accessibilityExpanded}
+      accessibilityHaspopup={accessibilityHaspopup}
+      accessibilityLabel={accessibilityLabel}
+      dataTestId={dataTestId}
       disabled={disabled}
+      fullWidth={fullWidth}
       name={name}
-      onBlur={handleBlur}
-      onClick={handleClick}
-      onMouseDown={handleMouseDown}
-      onMouseUp={handleMouseUp}
-      onTouchCancel={handleTouchCancel}
-      onTouchEnd={handleTouchEnd}
-      onTouchMove={handleTouchMove}
-      onTouchStart={handleTouchStart}
-      ref={innerRef}
-      tabIndex={disabled ? null : tabIndex}
-      type="button"
-    >
-      <div className={childrenDivClasses} style={compressStyle || undefined}>
-        {iconEnd ? (
-          <InternalButtonContent
-            text={buttonText}
-            textColor={textColor}
-            icon={iconEnd}
-            size={size}
-          />
-        ) : (
-          buttonText
-        )}
-      </div>
-    </button>
+      // $FlowExpectedError[incompatible-type]
+      ref={ref}
+      tabIndex={tabIndex}
+      selected={selected}
+      size={size}
+      text={text}
+      color={color}
+      iconEnd={iconEnd}
+      onClick={onClick}
+    />
   );
 });
