+++ packages/gestalt/src/Button/Button.jsdom.test.js
@@ -9,11 +9,7 @@ describe('Button', () => {
       [
         {|
           dangerouslyDisableOnNavigation: () => void,
-          event:
-            | SyntheticMouseEvent<HTMLButtonElement>
-            | SyntheticMouseEvent<HTMLAnchorElement>
-            | SyntheticKeyboardEvent<HTMLAnchorElement>
-            | SyntheticKeyboardEvent<HTMLButtonElement>,
+          event: SyntheticMouseEvent<HTMLButtonElement> | SyntheticKeyboardEvent<HTMLButtonElement>,
         |},
       ],
       void,
@@ -24,71 +20,12 @@ describe('Button', () => {
     expect(mockOnClick).toHaveBeenCalled();
   });
 
-  it('renders a submit button and forwards a ref to the innermost <button> element', () => {
-    const ref = createRef<HTMLButtonElement | HTMLAnchorElement>();
-    render(<Button type="submit" text="test" ref={ref} />);
-    expect(ref.current instanceof HTMLButtonElement).toEqual(true);
-    expect(ref.current?.type).toEqual('submit');
-  });
-
   it('renders a default button with sequential keyboard navigation and forwards a ref to the innermost <button> element', () => {
     const ref = createRef<HTMLButtonElement | HTMLAnchorElement>();
     render(<Button text="test" ref={ref} />);
     expect(ref.current instanceof HTMLButtonElement).toEqual(true);
     expect(ref.current?.type).toEqual('button');
     expect(ref.current instanceof HTMLButtonElement && ref.current?.tabIndex).toEqual(0);
-  });
-
-  it('renders a link button with correct new tab announcement with and without accessibilityLabel', () => {
-    render(
-      <Button
-        iconEnd="visit"
-        size="lg"
-        text="Visit Pinterest"
-        role="link"
-        rel="nofollow"
-        target="blank"
-        href="#"
-      />,
-    );
-    expect(
-      screen.getByText('Visit Pinterest', {
-        exact: true,
-      }),
-    ).toBeVisible();
-
-    expect(
-      screen.getByText('; Opens a new tab', {
-        exact: true,
-      }),
-    ).toBeVisible();
-
-    render(
-      <Button
-        accessibilityLabel="Visit Pinterest"
-        iconEnd="visit"
-        size="lg"
-        text="Visit Pinterest"
-        role="link"
-        rel="nofollow"
-        target="blank"
-        href="#"
-      />,
-    );
-
-    expect(screen.getByLabelText('Visit Pinterest; Opens a new tab')).toBeVisible();
-  });
-
-  it('renders a link button with sequential keyboard navigation and forwards a ref to the innermost <a> element', () => {
-    const ref = createRef<HTMLButtonElement | HTMLAnchorElement>();
-    render(
-      <Button text="test" role="link" href="http://www.pinterest.com" ref={ref} target="blank" />,
-    );
-    expect(ref.current instanceof HTMLAnchorElement).toEqual(true);
-    expect(ref.current instanceof HTMLAnchorElement && ref.current?.href).toEqual(
-      'http://www.pinterest.com/',
-    );
-    expect(ref.current instanceof HTMLAnchorElement && ref.current?.tabIndex).toEqual(0);
   });
 
   it('renders a disabled button', () => {
@@ -98,22 +35,6 @@ describe('Button', () => {
     expect(ref.current instanceof HTMLButtonElement && ref.current?.disabled).toEqual(true);
   });
 
-  it('renders a disabled link button', () => {
-    const ref = createRef<HTMLButtonElement | HTMLAnchorElement>();
-    render(
-      <Button
-        text="test"
-        role="link"
-        href="http://www.pinterest.com"
-        disabled
-        ref={ref}
-        target="blank"
-      />,
-    );
-    expect(ref.current instanceof HTMLAnchorElement).toEqual(true);
-    expect(ref.current instanceof HTMLAnchorElement && ref.current?.href).toEqual('');
-  });
-
   it('renders a button removed from sequential keyboard navigation via tabIndex', () => {
     const ref = createRef<HTMLButtonElement | HTMLAnchorElement>();
     render(<Button text="test" ref={ref} tabIndex={-1} />);
@@ -121,36 +42,9 @@ describe('Button', () => {
     expect(ref.current instanceof HTMLButtonElement && ref.current?.tabIndex).toEqual(-1);
   });
 
-  it('renders a link button removed from sequential keyboard navigation via tabIndex', () => {
-    const ref = createRef<HTMLButtonElement | HTMLAnchorElement>();
-    render(
-      <Button
-        text="test"
-        role="link"
-        href="http://www.pinterest.com"
-        ref={ref}
-        tabIndex={-1}
-        target="blank"
-      />,
-    );
-    expect(ref.current instanceof HTMLAnchorElement).toEqual(true);
-    expect(ref.current instanceof HTMLAnchorElement && ref.current?.tabIndex).toEqual(-1);
-  });
-
   it('renders with data-test-id', () => {
     const TEST_ID = 'button-test-123';
-    render(
-      <Button
-        dataTestId={TEST_ID}
-        iconEnd="visit"
-        size="lg"
-        text="Visit Pinterest"
-        role="link"
-        rel="nofollow"
-        target="blank"
-        href="#"
-      />,
-    );
+    render(<Button dataTestId={TEST_ID} iconEnd="visit" size="lg" text="Visit Pinterest" />);
     expect(
       screen.getByTestId(TEST_ID, {
         exact: true,
